+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Dev
+
+  * Remove `A.Array`
+
 ## v0.1.2 (2020-10-22)
 
 ### Enhancements