+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Dev
+## v0.2.0 (2020-10-25)
 
 ### Enhancements